requirements.txt
jupyter
gradio
google-genai